renaming persist package and interface
