+++ cmd/web/persist/interface.go
@@ -1,4 +1,4 @@
-package persister
+package persist
 
 import (
 	"context"
@@ -9,7 +9,7 @@ import (
 
 type PersistCallbackFn func(d hitlist.Domain, r hitlist.Recipient, vr validator.Result) error
 
-type Persist interface {
+type Persister interface {
 
 	// Store stores the parts and vr. The implementation decides what key to use, although it should use a similar one
 	// use to restore data using the Get or Range implementations
